Implemented Collisions for Bunkers
Made changes to the BunkerController header file by defining the BunkerController class as a class which implements the ICollider interface and declared and defined the interface's pure virtual methods in a way where bunker is destroyed on collision with SUBZERO enemy or with Torpedo bullet type and the bunker image sprite is returned in the header and source files.

Declared and defined function to get Bunker sprite in the BunkerView header and source files.

Declared and defined vector of ICollider pointers to instances of BunkerController class to store the instances to be deleted and function to delete the instances in the vector in the ElementService header and source files.

Made personal changes by:

1. Implementing the OutscalBomb powerup in a way that adds all the enemies wiped out by the bomb to the enemies killed by player instead of just wiping out enemies.

2. Implementing ThunderSnake enemy type by creating the Controller header and source files and for the time being implementing the same movement as Zapper and calling the spawn bullet function by sending Torpedo as bullet type in the source file.

+++ Space-Invaders/Source/Enemy/EnemyController.cpp
@@ -5,6 +5,7 @@
 #include "C:\Users\sidpa\Documents\GitHub\Space-Invaders\Space-Invaders\Header\ServiceLocator.h"
 #include "C:\Users\sidpa\Documents\GitHub\Space-Invaders\Space-Invaders\Header\Bullets\BulletController.h"
 #include "C:\Users\sidpa\Documents\GitHub\Space-Invaders\Space-Invaders\Header\Player\PlayerController.h"
+#include "C:\Users\sidpa\Documents\GitHub\Space-Invaders\Space-Invaders\Header\Elements\Bunker\BunkerController.h"
 #include "C:\Users\sidpa\Documents\GitHub\Space-Invaders\Space-Invaders\Header\Entities\EntityConfig.h"
 
 
@@ -12,6 +13,7 @@ namespace Enemy
 {
 	using namespace Global;
 	using namespace Entity;
+	using namespace Element::Bunker;
 
 	EnemyController::EnemyController(EnemyType type)
 	{
@@ -114,12 +116,22 @@ namespace Enemy
 			return;
 			
 		}
+
+		//If SUBZERO enemy has collided with bunker
+		BunkerController* bunker_controller = dynamic_cast<BunkerController*>(other_collider);
+		if (bunker_controller && getEnemyType() == EnemyType::SUBZERO) 
+		{
+			ServiceLocator::getInstance()->getEnemyService()->destroyEnemy(this);
+			return;
+		}
+			
 	}
 
 	void EnemyController::destroy()
 	{
-		ServiceLocator::getInstance()->getPlayerService()->increaseEnemiesKilled(1); 
-		ServiceLocator::getInstance()->getEnemyService()->destroyEnemy(this);
+		ServiceLocator::getInstance()->getPlayerService()->increaseEnemiesKilled(1);
+		ServiceLocator::getInstance()->getEnemyService()->destroyEnemy(this); 
+		
 	}
 
 	void EnemyController::render()
